feat: Add Models
SettingMeta model added

Ref: #9
