+++ .vscode/settings.json
@@ -7,7 +7,7 @@
   "editor.guides.indentation": true,
   "editor.renderWhitespace": "boundary",
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": "explicit"
   },
   "eslint.validate": [
     "javascript"
