+++ service/models.registry.js
@@ -1,3 +1,9 @@
-const modelsRegistry = {};
+const SettingMeta = require("./models/SettingMeta.model");
+const modelsRegistry = {
+    "SettingMeta":{
+        database: "application",
+        model: SettingMeta
+    }
+};
 
 exports.modelsRegistry = modelsRegistry;
